requirements.txt
django==4.1.7
django-ckeditor==6.5.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
phonenumbers==8.13.7
pre-commit==3.1.1

pandas~=1.5.3
Pillow~=9.4.0